tweaked gui

+++ client/GlassPreferencesBridge.cs
@@ -29,12 +29,15 @@ function GlassPrefGroup::cleanup() {
   GlassServerControlC.requested = false;
   GlassServerControlC::setTab(2);
   GlassPrefGroup.downloaded = 0;
-  for(%i = 0; %i < GlassPrefGroup.getCount(); %i++) {
-    %cat = GlassPrefGroup.getObject(%i);
-    %cat.deleteAll();
-  }
-  GlassPrefGroup.deleteAll();
-	GlassPrefGroup.delete();
+
+	if(isObject(GlassPrefGroup)) {
+	  for(%i = 0; %i < GlassPrefGroup.getCount(); %i++) {
+	    %cat = GlassPrefGroup.getObject(%i);
+	    %cat.deleteAll();
+	  }
+	  GlassPrefGroup.deleteAll();
+		GlassPrefGroup.delete();
+	}
 
 	new ScriptGroup(GlassPrefGroup);
 	$ServerInfo::PrefVersion = "";
